refactor: Rename add_review folder to write_review
